LayoutTests/imported/w3c/web-platform-tests/css/css-cascade/scope-import-implicit.tentative-expected.txt

Inside
Outside

FAIL Scope-imported rule applies within implicit scope assert_equals: expected "1" but got ""
PASS Scope-imported rule does not apply outside of implicit scope
